Physicalファイル分割
