+++ Game/KANJI/src/battle/physics/PhysicalWorldManager.hpp
@@ -5,48 +5,10 @@
 #include "CharaPhysicsParameters.hpp"
 
 namespace kanji {
-namespace chara {
-class IParameterizedCharacter;
-}
 namespace battle {
 class BattlePlayer;
-
-class PhysicalCharacter {
-public: // static_const/enum
-public: // static
-public: // public function
-    void update();
-    void drawLegacy() const; // 近いうちに消す
-private: // field
-    dx::di::PlayerId m_pid;
-    P2Body m_body;
-    const std::shared_ptr<chara::IParameterizedCharacter> m_status;
-    std::shared_ptr<param::CharaPhysics> m_param;
-private: // private function
-public: // ctor/dtor
-    PhysicalCharacter(
-        s3d::P2World* world,
-        dx::di::PlayerId pid,
-        const std::shared_ptr<chara::IParameterizedCharacter>& status,
-        const std::shared_ptr<param::CharaPhysics>& param);
-};
-
-
-class PhysicalStage {
-public: // static_const/enum
-public: // static
-public: // public function
-    void drawLegacy() const; // 近いうちに消す
-private: // field
-    const P2Body m_floor;
-    const P2Body m_ceiling;
-    const P2Body m_wall_left;
-    const P2Body m_wall_right;
-private: // private function
-public: // ctor/dtor
-    PhysicalStage(s3d::P2World* world, const std::shared_ptr<param::CharaPhysics>& param);
-};
-
+class PhysicalCharacter;
+class PhysicalStage;
 
 class PhysicalWorldManager {
 public: // static_const/enum
